requirements.txt
asgiref==3.3.1
Django==3.0.11
djangorestframework==3.12.2
Pillow==8.1.0
pytesseract==0.3.7
pytz==2020.5
sqlparse==0.4.1
gunicorn==20.0.4
PyMySQL==0.10.0